+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Mirror Globus directory to local cache.
 - Fix `--type_metadata` so it still works without a submission directory.
 - Add `--optional_fields` to temporarily ignore the given fields.
+- Add `--ignore_files` to ignore particular top-level files.
 - Ignore dot-files. No command-line option to enable stricter validation, for now.
 ### Changed
 - Make the ATACseq validation more flexible.